Refactor: Refactoring codes
