+++ admin/js/toast.js
@@ -1,4 +1,4 @@
-import { store } from '../../js/store.js'
+import { store } from '../../util/store.js';
 
 export function toast(message, state) {
   let main = '';
