--- js/store.js
@@ -1,8 +1,0 @@
-import dotenv from 'dotenv'
-
-export const store = {
-  selector: function $(selector) {
-    return document.querySelector(selector)
-  },
-  token: localStorage.accessToken
-}
