+++ js/userInfo.js
@@ -1,5 +1,5 @@
 import { accountLookUp, addAccount, editUser, accountCharge, cancelAccount } from "./requests.js";
-import { store } from './store.js'
+import { store } from '../util/store.js'
 
 // 회원 정보 수정 핸들러
 export function editUserInfo() {
